테스트 코드 리팩토링: EmailServiceTest
